requirements.txt
absl-py==2.3.0
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
gunicorn==23.0.0
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.1
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.8
pycparser==2.22
pyFirmata==1.1.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-pptx==1.0.2
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
tifffile==2024.8.30
typing_extensions==4.14.0
Werkzeug==3.1.3
xlsxwriter==3.2.5
zipp==3.23.0
